Update write_sequence API to take a uint64_t vector.
Includes a blind attempt to update waveforms.cpp

+++ src/lib/APS2.h
@@ -69,7 +69,7 @@ public:
 
 	int set_run_mode(const RUN_MODE &);
 
-	int write_sequence(const vector<uint32_t> &);
+	int write_sequence(const vector<uint64_t> &);
 	int clear_channel_data();
 
 	int load_sequence_file(const string &);
